don't amalgamate entry or import funcs during compilation
Import function addresses don't need to be known during compilation: they
are provided as a vector, mapped just before the text section.  The trap
handler is also part of the import vector.

The default (zero value) compile.CodeConfig cannot be used to access a
custom entry function, but the generated code will still invoke the
module's start function if there is one.  A simple ObjectMapper
implementation needs to be used to pick up the entry function address.

Below is an ObjectMapper implementation which is like object.FuncMap, but
stores only interesting (entry/export) function addresses.  Something
like this can be added to the API later if there's a need for it.

    // ExportMap implements ObjectMapper.  It stores addresses of the
    // non-import functions which were specified by the creator.
    type ExportMap struct {
        FuncAddrs map[uint32]int32

        funcIndex uint32
    }

    func NewExportMap(funcIndexes ...uint32) (m *ExportMap) {
        m = &ExportMap{
            FuncAddrs: make(map[uint32]int32),
        }
        for _, index := range funcIndexes {
            m.FuncAddrs[index] = 0
        }
        return
    }

    func (m *ExportMap) InitObjectMap(int, int) {
        m.funcIndex = 0
    }

    func (m *ExportMap) PutImportFuncAddr(int32) {
        m.funcIndex++
    }

    func (m *ExportMap) PutFuncAddr(addr int32) {
        if _, specified := m.FuncAddrs[m.funcIndex]; specified {
            m.FuncAddrs[m.funcIndex] = addr
        }
        m.funcIndex++
    }

    func (*ExportMap) PutCallSite(int32, int32) {}
    func (*ExportMap) PutInsnAddr(int32)        {}
    func (*ExportMap) PutDataBlock(int32, int)  {}