use intermediate wasm library for imports
The intermediate library is a simple wasm module.  It can import and call
functions only from the import vector.  It can't have data or globals.

Module.AsLibrary() returns a Library object, which is then used by
program modules.  Program modules can only import functions from the
library module.

Program's import functions get materialized in the program, by using code
from the library.  The vector calls used by the library code get
inlined, so trampoline functions not needed.

The benefit is that complexity and memory access can be moved to safe
wasm code.  The library code is trusted: secondary input validation by
machine code is not needed.

The downside is that the library function implementations add call and
trap sites to the object map, so their internal structure becomes part of
the portable ABI of the program.  The code can be changed as long as the
count and semantics of the call sites are backward- and forward-
compatible.