Redesigned library function resume mechanism
Library functions have only one suspension check, at the start of the
function.  Therefore the number and layout of call sites is not a
portability concern.

If a runtime function suspends itself and is resumed (portably), the whole
library function call needs to be rewound, and its argument values get
restored.  If library function needs to communicate state changes across
suspend/resume, it needs to use a memory buffer (argument).